fix: avoid use of some portion of STL in cuda
Since some of the standard template library header uses SIMD intrinsics
(depends on the implementation) that cannot be recognized by nvcc, we
may not be able to use it in *.cu file. `cuda_assert` may be called by
functions that should be compiled as cuda source code. To avoid error in
nvcc, remove most of the header from cuda_assert.