refactor: split world implementation and interface
Some of the STL implementation or C++ libraries are not compatible with
nvcc. Also, some of the cuda implementation are not compatible with gcc.
To avoid this conflict, we need to split cuda implementation
(including thrust::device_vector) from C++ header file. Here world_base
is defined to make interface and implement world in .cu file. .cu files
can share world implementation so it can use dynamic_cast to get world.